chore(notification): migrate :feature:notification module to kmp
