+++ feature/notification/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id(ThunderbirdPlugins.Library.android)
+    id(ThunderbirdPlugins.Library.kmp)
 }
 
 android {
