+++ feature/notification/src/androidMain/kotlin/net/thunderbird/feature/notification/NotificationLight.android.kt
@@ -1,0 +1,6 @@
+package net.thunderbird.feature.notification
+
+import android.app.Notification
+
+internal actual val NotificationLight.defaultColorInt: Int
+    get() = Notification.COLOR_DEFAULT
