+++ feature/notification/src/jvmMain/kotlin/net/thunderbird/feature/notification/NotificationLight.jvm.kt
@@ -1,0 +1,4 @@
+package net.thunderbird.feature.notification
+
+internal actual val NotificationLight.defaultColorInt: Int
+    get() = 0x00000000.toArgb()
